third_party/blink/web_tests/platform/generic/http/tests/security/mixedContent/insecure-script-through-redirection-expected.txt

CONSOLE ERROR: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-script-through-redirection.html' was loaded over HTTPS, but requested an insecure script 'http://example.test:8000/security/mixedContent/resources/script.js'. This request has been blocked; the content must be served over HTTPS.
This test opens a window that loads an insecure script through redirection. We should block it.
